Add MailerFactory service
In this commit I copied and slightly modified functionality used
for creating Mailer instance based on config from the Mailer to an extra class.

The functionality still remains also in Mailer and will be removed later.
I also created test based on the original test for Mailer and added test case
testing it creates specific MailerMethods.
[MAILPOET-4115]