Update cookie check in acceptance tests for PHP8.2
In PHP 8.2 (or perhaps in 8.1, as we updated tests from 8.0),
there was a change that, when fetching from DB started returning
integers as integers, not strings as before.

This was causing the test to fail with PHP 8.2.
We were getting an error: 'Failed asserting that 2 is of type "string".'
We still support older PHP, so I fixed the issue by casting it back to a string.

[MAILPOET-4620]