Fix issue with CSS variables in the style panel
Some components in the style setting (e.g., SpacingSizesControl) don't work
with the CSS variables when present in valid CSS form (e.g., var(--wp--preset--spacing--10)).
To prevent issues, we need to transform CSS variables to the shortened variant.
The core uses more sophisticated functions for that, see
https://github.com/WordPress/gutenberg/blob/f2ba0fdb4c7714e516f7e9045ac53e59401d5b82/packages/block-editor/src/components/global-styles/utils.js#L307-L358
The functionality is private, so I've created a simpler version.
[MAILPOET-6335]