Offset opt-in reruns from sending reruns
In automations, the send-email action can be rerun to wait for
subscriber opt-in, but also to wait for the the email scheduler (retry
sending). We use a simple counter to know how many times the action was
retried.

Both "reruns" have their own retry schedule however. Once opt-in has
been acquired, we must not mix the opt-in retry count, with the sending
retry count.