Refactor installation of specific WP version in test env
This commit changes how we install a specific WP version when required
for tests via the WORDPRESS_VERSION env variable.
In the previous approach, we installed the original version from the image
and then updated to the required version. We had some issues as after the
update, there might be some leftover background tasks, etc. that may
interfere with the tests.
In the new approach, we download the required version
prior to the installation.
[MAILPOET-6501]