Use project-specific GitHub client when fetching the latest release
When only free plugin client was used, once the free plugin was released, the release script for premium plugin used the just released version of the free and bumped the minor version again - but that didn't exist in the premium repo.

This way, each plugin checks its own releases. This shouldn't affect determining if patch or minor version should be bumped, because premium release always updates minor, and the free release checks the presence of the `release` branch in the premium repo.

STOMAIL-7398