Add migration to re-subscribe subscribers unsubscribed by bots
In the migration, we consider subscribers who made three and more clicks
and unsubscribed within 4 seconds before and 4 seconds after unsubscribe
as unsubscribed by a bot.

From research of affected sites, it seems that the issue started in late March or early April.
So we restrict the date to the start of March.
STOMAIL-7399