Add default value for re-engagement emails
This value is taken from the "Inactive subscribers" value. E.g. when the subscribers become inactive after 6 month, the value in re-engagement emails is 5 months.
But when the "Inactive subscribers" feature is turned off, there is no default value.

This PR changes it to 11 months, which would be used when the default "Inactive subscribers" value of 12 months is used

STOMAIL-7285